Refactor regex evaluation method across multiple linter rules
