+++ server/src/linter/rules/ambiguous/AM04.ts
@@ -42,7 +42,7 @@ export class ColumnCount extends Rule<string> {
     }
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
     return null;
