+++ server/src/linter/rules/capitalisation/CP04.ts
@@ -50,7 +50,7 @@ export class Literals extends Rule<string> {
     }
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
     return null;
