+++ server/src/linter/rules/convention/CV02.ts
@@ -44,7 +44,7 @@ export class Coalesce extends Rule<string> {
     }
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
     return null;
