+++ server/src/linter/rules/convention/CV08.ts
@@ -44,7 +44,7 @@ export class LeftJoin extends Rule<string> {
     }
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
     return null;
