+++ server/src/linter/rules/layout/LT01.ts
@@ -59,7 +59,7 @@ export class TrailingSpaces extends Rule<string> {
     }
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
     return null;
