+++ server/src/linter/rules/layout/LT10.ts
@@ -51,7 +51,7 @@ export class SelectModifiers extends Rule<string> {
     }
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
     return null;
