+++ server/src/linter/rules/layout/LT11.ts
@@ -52,7 +52,7 @@ export class UnionCheck extends Rule<string> {
     }
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
     return null;
