+++ server/src/linter/rules/layout/LT12.ts
@@ -39,7 +39,7 @@ export class EndofFile extends Rule<string> {
 		}
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
 		return null;
