+++ server/src/linter/rules/structure/ST01.ts
@@ -51,7 +51,7 @@ export class ElseNull extends Rule<string> {
     }
 
     if (this.pattern.test(test)) {
-      return this.evaluateMultiRegexTest(test, documentUri);
+      return this.evaluateRegexPatterns(test, documentUri);
     }
 
     return null;
